Ai2c Apps/Amap-backend-dev/requirements.txt
Django==4.2.25
mssql-django==1.6
python-dotenv==1.1.0
pandas==2.2.3
black==25.1.0
azure-storage-blob==12.25.1
azure-monitor-opentelemetry==1.6.10
XlsxWriter==3.2.0
django-simple-history==3.8.0
django-polymorphic==4.1.0
django-ninja==1.4.2
django-cors-headers==4.7.0
pillow==11.0.0
tqdm==4.67.1
django-csp==4.0
gunicorn==23.0.0
azure-identity==1.23.0
defusedxml==0.7.1
rapidfuzz>=3.0.0
pydantic==2.11.10